Add a new MockHttpTestCase class which behaves the same as the base MockHttp class, but you can also use assert methods if you class inherits from this one.
git-svn-id: https://svn.apache.org/repos/asf/incubator/libcloud/trunk@1125696 13f79535-47bb-0310-9956-ffa450edef68
