+++ test/loadbalancer/test_gogrid.py
@@ -7,7 +7,7 @@ from urlparse import urlparse, parse_qsl
 from libcloud.loadbalancer.base import LoadBalancer, Member, Algorithm
 from libcloud.loadbalancer.drivers.gogrid import GoGridLBDriver
 
-from test import MockHttp, MockRawResponse
+from test import MockHttpTestCase
 from test.file_fixtures import LoadBalancerFileFixtures
 
 class GoGridTests(unittest.TestCase):
@@ -84,19 +84,8 @@ class GoGridTests(unittest.TestCase):
 
         self.assertTrue(ret)
 
-class GoGridLBMockHttp(MockHttp, unittest.TestCase):
+class GoGridLBMockHttp(MockHttpTestCase):
     fixtures = LoadBalancerFileFixtures('gogrid')
-
-    def __init__(self, *args, **kwargs):
-        unittest.TestCase.__init__(self)
-
-        if kwargs.get('host', None) and kwargs.get('port', None):
-            MockHttp.__init__(self, *args, **kwargs)
-
-    def runTest(self):
-        # @TODO: Add a new base MockHttpTestCase class and add this method and
-        # constructor there
-        pass
 
     def _api_grid_loadbalancer_list(self, method, url, body, headers):
         body = self.fixtures.load('loadbalancer_list.json')
